+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 - Bugfix MN180N18 (bootloader 2.10.8) update not working (`ulTaskNotifyTakeIndexed` without time to wait not reliable) ([#52](https://github.com/OpenRemise/Firmware/issues/52))
+- Bugfix turn on Bug LED during DCC service mode ([#54](https://github.com/OpenRemise/Firmware/issues/54))
 
 ## 0.2.0
 - Support static IP ([#49](https://github.com/OpenRemise/Firmware/issues/49))